New Intent Parser: LookupIntentParser
What
--
This introduces a new dictionary-based deterministic intent parser.

The parser uses a hashmap to map normalized utterances to their
expected intent and slots.

Symbol tables are used to store the intents names and slot names
and this reduces the amount of information stored in the dictionary
values.

Why
--
This implementation is compact enough to store a huge number of
utterances and their expected outputs. It also out performs the
regex implementation as dict lookups are faster.

**NOTE**: The space usage can even be reduced further by hashing the
dictionary keys. (_this is a future plan once this impl is stabilized_)