Remove normalization of confidence scores in intent classification
The logreg intent classifier used to have a specific logic when used
with an intents filter: the intent classification scores were
renormalized to sum to 1.0.

This could lead to unexpected behaviors, especially when all the intents
in the intents filter are associated to very low confidence scores. In such
cases, the renormalization would significantly increase the scores.