Lookup Intent Parser (#759)
* New Intent Parser: LookupIntentParser

What
--
This introduces a new dictionary-based deterministic intent parser.

The parser uses a hashmap to map normalized utterances to their
expected intent and slots.

Symbol tables are used to store the intents names and slot names
and this reduces the amount of information stored in the dictionary
values.

Why
--
This implementation is compact enough to store a huge number of
utterances and their expected outputs. It also out performs the
regex implementation as dict lookups are faster.

**NOTE**: The space usage can even be reduced further by hashing the
dictionary keys. (_this is a future plan once this impl is stabilized_)

* add parser module + code linting

* remove unused attr

* fix: refactor based on review

* simplify implementation for consistency and simplicity:
    - avoid calculating range mappings
    - refactor text preprocessing and avoid code duplication
* fix bug
    - fix key collision handling
* fix up the docs

* fix preprocessing text

* fix text preprocessing

* linting fixes

* linting fixes

* remove unnecessary logic

* fixes from review

* lint fixes

* Add key hashing to Lookup Intent Parser for perf

Use hashes as dictionary keys instead of actual strings. This
reduces the size of the model.

Add utility function to handle key deserialization from JSON.

* Make dict iteration lazy + remove parens

* Fix failing test

* Apply fixes made on DeterministicIntentParser to LookupIntentParser

* Generate entities combinations to cover ambivalent words

* Fix non deterministic test

* Fix get_intents method of LookupIntentParser

* Fix processing order in DeterministicIntentParser

* Improve test

* Fix randomness in lookup intent parser

* Remove verbose mode in appveyor

* Fix tests